+++ ispyb/connector/mysqlsp/main.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import, division, print_function
-
 import threading
 import traceback
 
@@ -97,12 +95,10 @@ class ISPyBMySQLSPConnector(ispyb.interface.connection.IF):
             try:
                 result_args = cursor.callproc(procname=procname, args=args)
             except DataError as e:
-                raise ReadWriteError(
-                    "DataError({0}): {1}".format(e.errno, traceback.format_exc())
-                )
+                raise ReadWriteError(f"DataError({e.errno}): {traceback.format_exc()}")
             except IntegrityError as e:
                 raise ReadWriteError(
-                    "IntegrityError({0}): {1}".format(e.errno, traceback.format_exc())
+                    f"IntegrityError({e.errno}): {traceback.format_exc()}"
                 )
             finally:
                 cursor.close()
@@ -115,15 +111,16 @@ class ISPyBMySQLSPConnector(ispyb.interface.connection.IF):
             try:
                 cursor.callproc(procname=procname, args=args)
             except DataError as e:
-                raise ReadWriteError(
-                    "DataError({0}): {1}".format(e.errno, traceback.format_exc())
-                )
+                raise ReadWriteError(f"DataError({e.errno}): {traceback.format_exc()}")
 
             result = []
             for recordset in cursor.stored_results():
                 if isinstance(cursor, mysql.connector.cursor.MySQLCursorDict):
                     for row in recordset:
-                        result.append(dict(list(zip(recordset.column_names, row))))
+                        if isinstance(row, dict):
+                            result.append(row)
+                        else:
+                            result.append(dict(list(zip(recordset.column_names, row))))
                 else:
                     result = recordset.fetchall()
 
@@ -141,9 +138,7 @@ class ISPyBMySQLSPConnector(ispyb.interface.connection.IF):
                     args,
                 )
             except DataError as e:
-                raise ReadWriteError(
-                    "DataError({0}): {1}".format(e.errno, traceback.format_exc())
-                )
+                raise ReadWriteError(f"DataError({e.errno}): {traceback.format_exc()}")
             result = None
             rs = cursor.fetchone()
             if len(rs) > 0:
@@ -162,12 +157,10 @@ class ISPyBMySQLSPConnector(ispyb.interface.connection.IF):
                     args,
                 )
             except DataError as e:
-                raise ReadWriteError(
-                    "DataError({0}): {1}".format(e.errno, traceback.format_exc())
-                )
+                raise ReadWriteError(f"DataError({e.errno}): {traceback.format_exc()}")
             except IntegrityError as e:
                 raise ReadWriteError(
-                    "IntegrityError({0}): {1}".format(e.errno, traceback.format_exc())
+                    f"IntegrityError({e.errno}): {traceback.format_exc()}"
                 )
             result = None
             rs = cursor.fetchone()
@@ -184,6 +177,4 @@ class ISPyBMySQLSPConnector(ispyb.interface.connection.IF):
         try:
             cursor.execute("SET ROLE %s" % role)
         except DatabaseError as e:
-            raise ISPyBException(
-                "DatabaseError({0}): {1}".format(e.errno, traceback.format_exc())
-            )
+            raise ISPyBException(f"DatabaseError({e.errno}): {traceback.format_exc()}")