+++ ispyb/model/__init__.py
@@ -1,7 +1,14 @@
-from __future__ import absolute_import, division, print_function
+import warnings
+
+warnings.warn(
+    "ispyb.model has been deprecated and will be removed in a future version. "
+    "You can use the SQLAlchemy models in ispyb.sqlalchemy instead",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 
-class DBCache(object):
+class DBCache:
     """A helper class with useful functions to manage caching of database
     requests.
     Subclasses must implement reload() which should store data to be cached
@@ -34,7 +41,7 @@ class DBCache(object):
         return hasattr(self, "_data_cache")
 
 
-class EncapsulatedValue(object):
+class EncapsulatedValue:
     """A helper class encapsulating another object and mostly behaving as that
     object. The property .value allows access to the original object.
     A list of magic methods is implemented to allow meaningful comparisons.