+++ tests/conftest.py
@@ -1,9 +1,8 @@
 # pytest configuration file
 
-from __future__ import absolute_import, division, print_function
 
 import os
-import ispyb
+import ispyb.sqlalchemy
 import pytest
 
 
@@ -39,3 +38,8 @@ def testconfig_ws():
             "No configuration file for websocket tests found. Skipping websocket tests"
         )
     return config_file
+
+
+@pytest.fixture
+def alchemy(testconfig):
+    return ispyb.sqlalchemy.session(testconfig)